Complete first draft of security rules.
The rules are 'done', and uploaded to Firebase. They pass Firebase's
validation, but I haven't tested them yet. According to [this][1]
thread, there is a way to use the Forge simulator in code, so I could
write a set of acceptance tests for the security rules. I plan to
investigate this further.

[1]: https://groups.google.com/forum/#!topic/firebase-talk/I5dpIophviE