Plan should gracefully handle unit deletions.
Right now, if a unit gets deleted that is being used in a plan, then
the plan just switches to use the primary unit by default.

In the future, I think I'd rather always have the delete button available,
but it will give you a message if the unit is being used somewhere.

(Same with categories)